backend/requirements.txt
requests
pandas
uvicorn
fastapi
gql
aiohttp
scipy
httpx
polyline
python-dotenv
overpy
